WIP/POC: GoAwayRecv handler for Client and Server
This is awfully complex and should be replaced by some other method of
handling arbitrary frames. Ideally one more general and simpler, perhaps
just a FnMut callback. We'll need to figure out how to expose such
frames to user code, especially non-standard frames. But, I wrote this
anyway as an exercise to explore what would be necessary to make a fully
featured handler just for GOAWAYs.

While multiple different tasks/threads are dealing with an H2 session
and streams from it, it may be necessary to do special handling for any
GOAWAY frames that are received from the peer.  In particular, a
SendRequest ought to be removed from consideration for use for opening
new streams before it is given the GOAWAY and prevents any new streams
from being created.

We need to be able to manage multiple CONNECT-proxy streams made through
an H2 Session. If the H2 server begins a graceful shutdown with a GOAWAY
with a Last-Stream-Id of MAX_INT, any streams opened roughly
concurrently to the receipt of the GOAWAY frame ought to still be valid.

So, before .await-ing or otherwise polling a Connection to its end, a
user may pop off a GoAwayRecv handler.  While this exists, it must also
be actively polled. When a GOAWAY is received, poll_go_away() will
produce a BlockingGoAway, which wraps a GoAway that may be inspected,
cloned, or whatever. When the BlockingGoAway itself is dropped, the
underlying H2 Connection will then handle the frame and continue working
as usual. If a user cares to observe what may be multiple GOAWAYs, this
same cycle can be repeated forever. If the GoAwayRecv is dropped, the
Connection will continue uninterrupted as further GOAWAYs are received.